Added airBrake to make movement more like the default Roblox humanoid system.
I added a airBrake constant. I felt that the chickynoid system wasn't feeling like the default Roblox movement system. I came to the conclusion that it felt like you lost control in the air when jumping and moving, which was because the character kept on moving even though the player had stopped moving.